rename sidebar-toggle to menu-toggle
