+++ src/scripts/react/layout/menu-toggle/MenuToggleActionCreators.js
@@ -1,16 +1,16 @@
 import dispatcher from '../../../Dispatcher';
-import { ActionTypes } from './SidebarToggleConstants';
+import { ActionTypes } from './MenuToggleConstants';
 
 export default {
   open: () => {
     dispatcher.handleViewAction({
-      type: ActionTypes.SIDEBAR_TOGGLE_OPEN
+      type: ActionTypes.MENU_TOGGLE_OPEN
     });
   },
 
   close: () => {
     dispatcher.handleViewAction({
-      type: ActionTypes.SIDEBAR_TOGGLE_CLOSE
+      type: ActionTypes.MENU_TOGGLE_CLOSE
     });
   }
 };
