+++ src/scripts/react/layout/menu-toggle/MenuToggleConstants.js
@@ -1,6 +1,6 @@
 import keyMirror from 'fbjs/lib/keyMirror';
 
 export const ActionTypes = keyMirror({
-  SIDEBAR_TOGGLE_OPEN: null,
-  SIDEBAR_TOGGLE_CLOSE: null
+  MENU_TOGGLE_OPEN: null,
+  MENU_TOGGLE_CLOSE: null
 });
