+++ src/styles/app.less
@@ -4,7 +4,7 @@
 @import "Guide";
 @import "Badges";
 @import "Label-backend";
-@import "Toggle";
+@import "MenuToggle";
 
 // imports 3rd party themes
 // lets reuse KBC theme variables
