+++ src/styles/theme-kbc-mobile-sidebar.less
@@ -1,4 +1,4 @@
-@media screen and (max-width: 768px) {
+@media (max-width: 768px) {
   body {
     padding-top: 120px;
   }
@@ -30,7 +30,7 @@
   .sidebar-offset-row {
     position: relative;
 
-    &.active {
+    &-open {
       .sidebar-offset {
         transform: translate3d(0, 0, 0);
       }
